Fix changing commits with non-ASCII characters
This fixes commits with non-ASCII characaters. Instead of using the char index
as byte index into the string, it tries to find the positions of the newline
characters.

This also fixes the case of signed commits, adding the padding directly after
"gpgsig " leads to git not finding the signature. Therefore it adds padding in
the line after `-----BEGIN PGP SIGNATURE-----`.