Support sha256-format git repositories
I'm not really sure why I added this, because sha256 git repositories are highly experimental and are used exceedingly rarely. That said, the final version ended up being simpler than expected now that const generics are available on Rust stable. In typical Rust fashion, it took many hours to get this to compile, and then it worked perfectly the first time I ran it.