Add support for a help flag
While lucky-commit has a pretty simple interface, because it operates
by side-effects it can be safer to check its behaviour.

Currently, to get the help it's necessary to be lucky enough to
mistakenly give it 2 args, thus failing to match any case. Otherwise,
it will panic if called from not-within-a-repository, and silently
rewrite the HEAD if invoked from one.