Print new hash on stdout on success
While it's currently possible to check the new value of HEAD, it
becomes an issue if support for converting non-HEAD commits (whether
on a ref or raw OIDs) becomes an option. While looking up the new
commit by prefix might be possible it's annoying (even more so on
lucky-commit where every commit is luckified).