support explicit ref or commit (not just HEAD)
Implicitly working with `HEAD` can be a chore when trying to use LC on
a bare repo or "headless" (scripted).

Add support for an optional 2nd parameter, which defaults to `HEAD`
but can be any ref or commit oid. If an oid, `update-ref` is skipped.

The skip heuristic is to `show-ref` the input, it'll fail on oids but
should succeed on every resolveable ref (which we know `ref_` is
because we passed it to `cat-file` and it succeeded).